implement mutation post photo
